dara-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
openai
httpx
google-cloud-texttospeech
